fix: finessed file routing server
