+++ lib/http-server.js
@@ -7,16 +7,16 @@ import { jsonParseSafe, errorResponse, successResponse } from './helpers.js';
 const method = { get: 'read', post: 'create', put: 'update', options: 'options' };
 
 const parseOptions = {
-  'text/plain': (buffer) => buffer.toString(),
-  'text/html': (buffer) => buffer.toString(),
+  'text/plain': (text) => text,
+  'text/html': (text) => text,
   'application/json': (json) => jsonParseSafe(json, {}),
-  'application/x-www-form-urlencoded': (buffer) => Object.fromEntries(new URLSearchParams(buffer.toString())),
+  'application/x-www-form-urlencoded': (text) => Object.fromEntries(new URLSearchParams(text)),
 };
 
 async function parseBody(req, parser) {
   const buffers = [];
   for await (const chunk of req) buffers.push(chunk);
-  const data = Buffer.concat(buffers);
+  const data = Buffer.concat(buffers).toString();
   return parser(data);
 }
 
@@ -48,6 +48,7 @@ export function createHttpServer(routing, port = 8000) {
     .listen(port, () => console.log('Server started on port: ', port));
 
   function gracefulShutdown() {
+    console.log('graceful shutdown');
     server.close((error) => {
       if (error) {
         console.error(error);
@@ -61,7 +62,7 @@ export function createHttpServer(routing, port = 8000) {
     socket.end('Bad Request');
   });
 
-  process.on('SIGTERM', async () => {
+  process.on('SIGINT', async () => {
     gracefulShutdown();
   });
 }
